+++ src/modules/cars/repositories/implementations/CategoriesRepository.ts
@@ -1,5 +1,5 @@
-import { Category } from "../model/Category";
-import { ICategoriesRepository } from "./ICategoriesRepository";
+import { Category } from "../../model/Category";
+import { ICategoriesRepository } from "../ICategoriesRepository";
 
 interface ICreateCategoryDTO {
   name: string;
